Update letter.html
changed the order of some things and the layout again, decided to change the font to monaco because I did not like the way arial was looking. I searched up how to do it. I did change a bit how the lines are ordered but overall more just visual and stylistics changed